rename artifact search to search
